Refactor plugin dir structure and add early-init.el
Refactor non melpa (personal) plugins to be self contained in a "plugins"
folder rather than the root directory. Additionally add the early-init.el
file, which loads prior to the init.el file, which is compiled from settings.org.
Include a statement in this new file to supress the warning that the core
library "cl" is technically deprecated, while melpa/elpa plugins work to
replace the cl library with the cl-libs successor library.